feat: merge useTouchZoom
