+++ src/hooks/useTouchZoom.ts
@@ -1,6 +1,10 @@
 import { useCallback, useRef } from 'react';
+import type { TransformAction, TransformType } from './useImageTransform';
 
-export default function useTouchZoom() {
+export default function useTouchZoom(
+  updateTransform: (newTransform: Partial<TransformType>, action: TransformAction) => void,
+  currentScale: number,
+) {
   const touchPointInfo = useRef({
     touchOne: 0,
     touchTwo: 0,
@@ -19,5 +23,49 @@ export default function useTouchZoom() {
     touchPointInfo.current.touchdown = true;
   }, []);
 
-  return { touchPointInfo: touchPointInfo.current, restTouchPoint, setTouchPoint };
+  const onTouchStart = useCallback(
+    (event: React.TouchEvent<HTMLImageElement>) => {
+      const { touches = [] } = event;
+      if (touches.length > 1) {
+        setTouchPoint(touches[0].pageY, touches[1].pageY);
+      }
+    },
+    [setTouchPoint],
+  );
+
+  const onTouchMove = (event: React.TouchEvent<HTMLImageElement>) => {
+    const { touches = [] } = event;
+    const { touchOne, touchTwo, touchdown } = touchPointInfo.current;
+
+    if (touchdown) {
+      const pageY_1 = touches[0]?.pageY;
+      const pageY_2 = touches[1]?.pageY;
+      let needChange = false;
+
+      if (Math.abs(touchOne - pageY_1) > 10 || Math.abs(touchTwo - pageY_2) > 10) {
+        needChange = true;
+      }
+
+      if (needChange) {
+        if (Math.abs(touchOne - touchTwo) > Math.abs(pageY_1 - pageY_2)) {
+          if (currentScale <= 1) return;
+          updateTransform(
+            { x: 0, y: 0, scale: currentScale - 0.3 < 1 ? 1 : currentScale - 0.3 },
+            'touchZoom',
+          );
+        } else {
+          updateTransform({ x: 0, y: 0, scale: currentScale + 0.2 }, 'touchZoom');
+        }
+
+        setTouchPoint(pageY_1, pageY_2);
+      }
+    }
+  };
+
+  return {
+    touchPointInfo: touchPointInfo.current,
+    onTouchRest: restTouchPoint,
+    onTouchStart,
+    onTouchMove,
+  };
 }
